LD密码现在有50%概率交换牌库。匹配密码不再区分大小写。编辑器新增黑名单，不会匹配到含有黑名单中的卡的卡组

+++ src/Cynthia.Card.Unity/src/Cynthia.Unity.Card/Assets/Script/EditorMenu/EditorInfo.cs
@@ -39,7 +39,7 @@ public class EditorInfo : MonoBehaviour
     private IDictionary<Faction, GameObject> _deckPrefabMap;
     private int _nowShow = -1;
     private string _showSearchMessage = "";
-    private bool isSpecial=false;
+    private bool isSpecial = false;
     //------------------------------------------------
     //DOTween动画
     public RectTransform ShowCardsTitle;
@@ -68,7 +68,7 @@ public class EditorInfo : MonoBehaviour
     private Faction _nowSwitchFaction = Faction.All;
     private string _nowSwitchLeaderId = null;
     //乱斗模式禁止阿瓦拉克，希拉德，蒂博尔，大锤，店店
-    private List<string> SpecialBanningList=new List<string>
+    private List<string> SpecialBanningList = new List<string>
             { "12032","12011","12012", "32007", "32005", "22009", "12041" };
     //------------------------------------------------
     //卡组编辑相关
@@ -147,11 +147,11 @@ public class EditorInfo : MonoBehaviour
             {
                 var card = Instantiate(EditorMenuCardPrefab).GetComponent<EditorUICoreCard>();
                 card.cardShowInfo.setCurrentCore(x, true);
-                var canAdd=0;
-                if(!isSpecial)
-                canAdd = (x.Group == Group.Copper ? 3 : 1);
-                else 
-                canAdd = ((x.Group == Group.Gold ||x.Group == Group.Copper)? 3 : 1);
+                var canAdd = 0;
+                if (!isSpecial)
+                    canAdd = (x.Group == Group.Copper ? 3 : 1);
+                else
+                    canAdd = ((x.Group == Group.Gold || x.Group == Group.Copper) ? 3 : 1);
                 card.Count = (canAdd - _nowEditorDeck.Deck.Where(c => c == x.CardId).Count());
                 card.transform.SetParent(EditorCardsContext, false);
             });
@@ -300,7 +300,7 @@ public class EditorInfo : MonoBehaviour
         {
             if (_deckPrefabMap == null) Start();
             var deck = Instantiate(_deckPrefabMap[GwentMap.CardMap[x.Leader].Faction]);
-            deck.GetComponent<DeckShowInfo>().SetDeckInfo(x.Name, x.IsBasicDeck()||x.IsSpecialDeck());
+            deck.GetComponent<DeckShowInfo>().SetDeckInfo(x.Name, x.IsBasicDeck() || x.IsSpecialDeck() || (x.IsBlacklist() && x.Id == "blacklist"));
             deck.GetComponent<EditorShowDeck>().Id = x.Id;
             deck.transform.SetParent(ShowDecksContext, false);
         });
@@ -313,19 +313,22 @@ public class EditorInfo : MonoBehaviour
 
     public async void ShowDeckRemoveClick(string Id)
     {
-        if (await _globalUIService.YNMessageBox("PopupWindow_DeleteDeckTitle",
-            string.Format(_translator.GetText("PopupWindow_DeleteDeckDesc"), _clientService.User.Decks.Single(x => x.Id == Id).Name)))
+        if (_nowEditorDeck.Id != "blacklist")
         {
-            if (!(await _clientService.RemoveDeck(Id)))
+            if (await _globalUIService.YNMessageBox("PopupWindow_DeleteDeckTitle",
+                string.Format(_translator.GetText("PopupWindow_DeleteDeckDesc"), _clientService.User.Decks.Single(x => x.Id == Id).Name)))
             {
-                await _globalUIService.YNMessageBox("PopupWindow_DeleteDeckErrorTitle",
-                    string.Format(_translator.GetText("PopupWindow_DeleteDeckErrorDesc"), _clientService.User.Decks.Single(x => x.Id == Id).Name));
-            }
-            else
-            {
-                var i = _clientService.User.Decks.Select((item, index) => (item, index)).Single(x => x.item.Id == Id).index;
-                _clientService.User.Decks.RemoveAt(i);
-                SetDeckList(_clientService.User.Decks);
+                if (!(await _clientService.RemoveDeck(Id)))
+                {
+                    await _globalUIService.YNMessageBox("PopupWindow_DeleteDeckErrorTitle",
+                        string.Format(_translator.GetText("PopupWindow_DeleteDeckErrorDesc"), _clientService.User.Decks.Single(x => x.Id == Id).Name));
+                }
+                else
+                {
+                    var i = _clientService.User.Decks.Select((item, index) => (item, index)).Single(x => x.item.Id == Id).index;
+                    _clientService.User.Decks.RemoveAt(i);
+                    SetDeckList(_clientService.User.Decks);
+                }
             }
         }
     }
@@ -335,7 +338,7 @@ public class EditorInfo : MonoBehaviour
         Debug.Log("点击了【" + _clientService.User.Decks.Single(x => x.Id == Id).Name + "】卡组的编辑按钮");
         var deck = _clientService.User.Decks.Single(x => x.Id == Id);
         _nowEditorDeck = deck;
-        isSpecial= (!deck.IsHalfBasicDeck())&&deck.IsHalfSpecialDeck();
+        isSpecial = (!deck.IsHalfBasicDeck()) && deck.IsHalfSpecialDeck();
         _nowSwitchLeaderId = deck.Leader;
         _nowSwitchFaction = GwentMap.CardMap[deck.Leader].Faction;
         //
@@ -348,32 +351,33 @@ public class EditorInfo : MonoBehaviour
     }
     //=============================================================================================================================
     //以上为展示卡牌相关的内容,以下为展示框选择相关内容
-    public void SwitchDeckClick(){
-         var decks =_nowEditorDeck.Deck;
-        isSpecial=!isSpecial;
-        if(!isSpecial)
+    public void SwitchDeckClick()
+    {
+        var decks = _nowEditorDeck.Deck;
+        isSpecial = !isSpecial;
+        if (!isSpecial)
         {
-            var gold=decks.Where(x =>GwentMap.CardMap[x].Group == Group.Gold).Distinct().ToList().Take(4).ToList();
-            var silver=decks.Where(x => GwentMap.CardMap[x].Group == Group.Silver).ToList();
-            var copper=decks.Where(x => GwentMap.CardMap[x].Group == Group.Copper).ToList();
+            var gold = decks.Where(x => GwentMap.CardMap[x].Group == Group.Gold).Distinct().ToList().Take(4).ToList();
+            var silver = decks.Where(x => GwentMap.CardMap[x].Group == Group.Silver).ToList();
+            var copper = decks.Where(x => GwentMap.CardMap[x].Group == Group.Copper).ToList();
             gold.AddRange(silver);
             gold.AddRange(copper);
-            _nowEditorDeck.Deck=gold;
+            _nowEditorDeck.Deck = gold;
         }
         else
         {
-            var gold=decks.Where(x =>  (!SpecialBanningList.Contains(GwentMap.CardMap[x].CardId))&&(GwentMap.CardMap[x].Group == Group.Gold)).ToList();
-            var silver=decks.Where(x => GwentMap.CardMap[x].Group == Group.Silver).ToList();
-            var copper=decks.Where(x => GwentMap.CardMap[x].Group == Group.Copper).Distinct().ToList();
+            var gold = decks.Where(x => (!SpecialBanningList.Contains(GwentMap.CardMap[x].CardId)) && (GwentMap.CardMap[x].Group == Group.Gold)).ToList();
+            var silver = decks.Where(x => GwentMap.CardMap[x].Group == Group.Silver).ToList();
+            var copper = decks.Where(x => GwentMap.CardMap[x].Group == Group.Copper).Distinct().ToList();
             gold.AddRange(silver);
             gold.AddRange(copper);
-            _nowEditorDeck.Deck=gold; 
+            _nowEditorDeck.Deck = gold;
         }
-        
+
         SetEditorDeck(_nowEditorDeck);
-    //=============================================================================================================================
-       AutoSetEditorCards();
-       } 
+        //=============================================================================================================================
+        AutoSetEditorCards();
+    }
 
     public void AddDeckClick()
     {   //点击新建按钮后
@@ -381,7 +385,7 @@ public class EditorInfo : MonoBehaviour
         {
             _globalUIService.YNMessageBox(_translator.GetText("PopupWindow_DeckLimitTitle"), _translator.GetText("PopupWindow_DeckLimitDesc"));
         }
-       else
+        else
         {
             DOTween.To(() => ShowCardsTitle.anchoredPosition, x => ShowCardsTitle.anchoredPosition = x,
                         new Vector2(0, 150), 0.5f);//收回Title
@@ -402,7 +406,7 @@ public class EditorInfo : MonoBehaviour
             _nowEditorDeck = null;
             DeckName.text = _translator.GetText("EditorMenu_DefaultDeckname");
         }
-    
+
     }
 
     public void SelectSwitchUICard(CardStatus card, bool isOver = true)
@@ -494,7 +498,9 @@ public class EditorInfo : MonoBehaviour
                 // }
                 // else
                 // {
-                _nowEditorDeck.Name = (DeckName.text == "" ? _translator.GetText("EditorMenu_DefaultDeckname") : DeckName.text);
+                if (_nowEditorDeck.Id != "blacklist")
+                    _nowEditorDeck.Name = (DeckName.text == "" ? _translator.GetText("EditorMenu_DefaultDeckname") : DeckName.text);
+
                 if (_clientService.User.Decks.Any(x => x.Id == (_nowEditorDeck.Id == null ? "" : _nowEditorDeck.Id)))
                 {
                     // if (await _globalUIService.YNMessageBox("是否修改卡组?", $"是否修改卡组 {DeckName.text}"))
@@ -606,30 +612,42 @@ public class EditorInfo : MonoBehaviour
     public void ClickEditorUICoreCard(CardStatus card)
     {//点击了显示卡牌  应该判断是否应该添加卡牌,如果可以,添加并且更新显示,否则跳出消息提醒
         var count = _nowEditorDeck.Deck.Where(x => x == card.CardId).Count();
-          if(!isSpecial)
-       {
-        if (!((card.Group == Group.Copper && count >= 3) ||
-            (card.Group != Group.Copper && count >= 1) ||
-            (_nowEditorDeck.Deck.Count >= 40) ||
-            (card.Group == Group.Silver && _nowEditorDeck.Deck.Where(x => x.CardInfo().Group == Group.Silver).Count() >= 6) ||
-            (card.Group == Group.Gold && _nowEditorDeck.Deck.Where(x => x.CardInfo().Group == Group.Gold).Count() >= 4)))
-        {   //如果超过上限,禁止加入卡牌
-            _nowEditorDeck.Deck.Add(card.CardId);
-            SetEditorDeck(_nowEditorDeck);
-            //**********************************************
-            var c = GetAllChilds<EditorUICoreCard>(EditorCardsContext).Where(x => x.cardShowInfo.CurrentCore.CardId == card.CardId);
-            c.ForAll(x => { x.Count--; });
+        if (_nowEditorDeck.Id == "blacklist")
+        {
+            if (!(count >= 1 || (_nowEditorDeck.Deck.Count >= 3)))
+            {   //如果超过上限,禁止加入卡牌
+                _nowEditorDeck.Deck.Add(card.CardId);
+                SetEditorDeck(_nowEditorDeck);
+                //**********************************************
+                var c = GetAllChilds<EditorUICoreCard>(EditorCardsContext).Where(x => x.cardShowInfo.CurrentCore.CardId == card.CardId);
+                c.ForAll(x => { x.Count--; });
+            }
+
         }
+        else if (!isSpecial)
+        {
+            if (!((card.Group == Group.Copper && count >= 3) ||
+                (card.Group != Group.Copper && count >= 1) ||
+                (_nowEditorDeck.Deck.Count >= 40) ||
+                (card.Group == Group.Silver && _nowEditorDeck.Deck.Where(x => x.CardInfo().Group == Group.Silver).Count() >= 6) ||
+                (card.Group == Group.Gold && _nowEditorDeck.Deck.Where(x => x.CardInfo().Group == Group.Gold).Count() >= 4)))
+            {   //如果超过上限,禁止加入卡牌
+                _nowEditorDeck.Deck.Add(card.CardId);
+                SetEditorDeck(_nowEditorDeck);
+                //**********************************************
+                var c = GetAllChilds<EditorUICoreCard>(EditorCardsContext).Where(x => x.cardShowInfo.CurrentCore.CardId == card.CardId);
+                c.ForAll(x => { x.Count--; });
+            }
         }
         else
         {
-            if(!((card.Group == Group.Silver && count >= 1) ||
-               ((card.Group == Group.Gold ||card.Group == Group.Copper)&& count >= 3)||
+            if (!((card.Group == Group.Silver && count >= 1) ||
+               ((card.Group == Group.Gold || card.Group == Group.Copper) && count >= 3) ||
                (_nowEditorDeck.Deck.Count >= 40) ||
                (card.Group == Group.Silver && _nowEditorDeck.Deck.Where(x => x.CardInfo().Group == Group.Silver).Count() >= 6) ||
                (card.Group == Group.Gold && _nowEditorDeck.Deck.Where(x => x.CardInfo().Group == Group.Gold).Count() >= 12)))
             {
-                _nowEditorDeck.Deck.Add(card.CardId);      
+                _nowEditorDeck.Deck.Add(card.CardId);
                 SetEditorDeck(_nowEditorDeck);
                 //**********************************************
                 var c = GetAllChilds<EditorUICoreCard>(EditorCardsContext).Where(x => x.cardShowInfo.CurrentCore.CardId == card.CardId);
@@ -663,7 +681,7 @@ public class EditorInfo : MonoBehaviour
         (
             //
             _cards
-            .Where(x => (!(isSpecial&&SpecialBanningList.Contains(x.CardInfo().CardId))&&((x.Faction == Faction.Neutral) || (x.Faction == _nowSwitchFaction))))
+            .Where(x => ((_nowEditorDeck.Id == "blacklist") || (!(isSpecial && SpecialBanningList.Contains(x.CardInfo().CardId)) && ((x.Faction == Faction.Neutral) || (x.Faction == _nowSwitchFaction)))))
             .Where(x => ((_editorSearchMessage == "") ? true :
                 (_translator.GetCardName(x.CardInfo().CardId).Contains(_editorSearchMessage, StringComparison.OrdinalIgnoreCase) ||
                  _translator.GetCardInfo(x.CardInfo().CardId).Contains(_editorSearchMessage, StringComparison.OrdinalIgnoreCase) ||
@@ -695,13 +713,15 @@ public class EditorInfo : MonoBehaviour
             card.transform.SetParent(EditorCListContext, false);
         });
         AllCount.text = _nowEditorDeck.Deck.Count().ToString();
-          AllCount.color =(deck.IsSpecialDeck()||deck.IsBasicDeck()) ? ClientGlobalInfo.NormalColor : ClientGlobalInfo.ErrorColor;
-        AllCountText.color = (deck.IsSpecialDeck()||deck.IsBasicDeck()) ? ClientGlobalInfo.NormalColor : ClientGlobalInfo.ErrorColor;
+        bool valid = (!deck.IsBlacklist() && (deck.IsSpecialDeck() || deck.IsBasicDeck())
+        || deck.IsBlacklist() && deck.Id == "blacklist");
+        AllCount.color = valid ? ClientGlobalInfo.NormalColor : ClientGlobalInfo.ErrorColor;
+        AllCountText.color = valid ? ClientGlobalInfo.NormalColor : ClientGlobalInfo.ErrorColor;
         CopperCount.text = $"{_nowEditorDeck.Deck.Where(x => GwentMap.CardMap[x].Group == Group.Copper).Count()}";
-         if(isSpecial)
-         GoldCount.text = $"{_nowEditorDeck.Deck.Where(x => GwentMap.CardMap[x].Group == Group.Gold).Count()}/12";
-        else  
-        GoldCount.text = $"{_nowEditorDeck.Deck.Where(x => GwentMap.CardMap[x].Group == Group.Gold).Count()}/4";
+        if (isSpecial)
+            GoldCount.text = $"{_nowEditorDeck.Deck.Where(x => GwentMap.CardMap[x].Group == Group.Gold).Count()}/12";
+        else
+            GoldCount.text = $"{_nowEditorDeck.Deck.Where(x => GwentMap.CardMap[x].Group == Group.Gold).Count()}/4";
         SilverCount.text = $"{_nowEditorDeck.Deck.Where(x => GwentMap.CardMap[x].Group == Group.Silver).Count()}/6";
         //*****************
         //等待补充？？？
@@ -744,7 +764,7 @@ public class EditorInfo : MonoBehaviour
         var deck = deckCode.DeCompressToDeck();
         deck.Name = string.IsNullOrWhiteSpace(name) ? _translator.GetText("EditorMenu_DefaultDeckname")
                                                     : (name.Length >= 20 ? name.Substring(0, 20) : name);
-        if (!(deck.IsBasicDeck()|| deck.IsSpecialDeck()))
+        if (!(deck.IsBasicDeck() || deck.IsSpecialDeck()))
         {
             await _globalUIService.YNMessageBox("PopupWindow_AddDeckErrorTitle",
                                                "Code: " + deckCode, "PopupWindow_OkButton", isOnlyYes: true);
